agregados todos los test principales, fallan 3
