+++ features/ahorcado.feature
@@ -1,0 +1,35 @@
+Feature: showing off behave
+  Scenario: Repetir letra usada
+    Given la palabra es PALABRA
+    When jugador intenta las letras B,M,B
+    Then se muestra pantalla usada
+
+  Scenario: Ganar el juego intentanto palabra entera
+    Given la palabra es PALABRA
+    When jugador ingresa PALABRA
+    Then se muestra pantalla ganaste
+  
+  Scenario: Reiniciar juego luego de perder
+    Given se muestra pantalla perdiste
+    When jugador clickea en reiniciar
+    Then se muestra pantalla empezar_juego
+
+  Scenario: Reiniciar juego luego de ganar
+    Given se muestra pantalla ganaste
+    When jugador clickea en Reiniciar
+    Then se muestra pantalla empezar_juego
+
+  Scenario: Perder el juego intentanto palabra entera
+    Given la palabra es PALABRA
+    When jugador ingresa MANZANA
+    Then se muestra pantalla perdiste
+
+  Scenario: Quedarse sin vidas
+    Given la palabra es PALABRA
+    When jugador intenta las letras E,O,U,I,A,C,M,G
+    Then se muestra pantalla perdiste
+
+  Scenario: Ganar completando letras
+    Given la palabra es PALABRA
+    When jugador intenta las letras A,T,P,L,U,R,B
+    Then se muestra pantalla ganaste
